+++ examples/create_datagrid2_page/main.go
@@ -22,6 +22,8 @@ import (
 	"os"
 	"strings"
 
+	"github.com/mendixlabs/mxcli/mdl/backend"
+	mprbackend "github.com/mendixlabs/mxcli/mdl/backend/mpr"
 	"github.com/mendixlabs/mxcli/mdl/executor"
 	"github.com/mendixlabs/mxcli/mdl/visitor"
 )
@@ -51,6 +53,7 @@ func main() {
 
 	// Create the MDL executor with stdout for output
 	exec := executor.New(os.Stdout)
+	exec.SetBackendFactory(func() backend.FullBackend { return mprbackend.New() })
 
 	// Define the MDL script to create a page with DataGrid2
 	// Note: Adjust module name, entity name, and attributes to match your project
