--- mdl/executor/bson_helpers.go
@@ -1,481 +1,0 @@
-// SPDX-License-Identifier: Apache-2.0
-
-package executor
-
-import (
-	"fmt"
-	"strings"
-
-	"go.mongodb.org/mongo-driver/bson"
-	"go.mongodb.org/mongo-driver/bson/primitive"
-
-	"github.com/mendixlabs/mxcli/mdl/bsonutil"
-	mdlerrors "github.com/mendixlabs/mxcli/mdl/errors"
-	"github.com/mendixlabs/mxcli/mdl/types"
-)
-
-// ============================================================================
-// bson.D helper functions for ordered document access
-// ============================================================================
-
-// dGet returns the value for a key in a bson.D, or nil if not found.
-func dGet(doc bson.D, key string) any {
-	for _, elem := range doc {
-		if elem.Key == key {
-			return elem.Value
-		}
-	}
-	return nil
-}
-
-// dGetDoc returns a nested bson.D field value, or nil.
-func dGetDoc(doc bson.D, key string) bson.D {
-	v := dGet(doc, key)
-	if d, ok := v.(bson.D); ok {
-		return d
-	}
-	return nil
-}
-
-// dGetString returns a string field value, or "".
-func dGetString(doc bson.D, key string) string {
-	v := dGet(doc, key)
-	if s, ok := v.(string); ok {
-		return s
-	}
-	return ""
-}
-
-// dSet sets a field value in a bson.D in place. If the key exists, it's updated
-// and returns true. If the key is not found, returns false.
-func dSet(doc bson.D, key string, value any) bool {
-	for i := range doc {
-		if doc[i].Key == key {
-			doc[i].Value = value
-			return true
-		}
-	}
-	return false
-}
-
-// dGetArrayElements extracts Mendix array elements from a bson.D field value.
-// Handles the int32 type marker at index 0.
-func dGetArrayElements(val any) []any {
-	arr := toBsonA(val)
-	if len(arr) == 0 {
-		return nil
-	}
-	if _, ok := arr[0].(int32); ok {
-		return arr[1:]
-	}
-	if _, ok := arr[0].(int); ok {
-		return arr[1:]
-	}
-	return arr
-}
-
-// toBsonA converts various BSON array types to []any.
-func toBsonA(v any) []any {
-	switch arr := v.(type) {
-	case bson.A:
-		return []any(arr)
-	case []any:
-		return arr
-	default:
-		return nil
-	}
-}
-
-// dSetArray sets a Mendix-style BSON array field, preserving the int32 marker.
-func dSetArray(doc bson.D, key string, elements []any) {
-	existing := toBsonA(dGet(doc, key))
-	var marker any
-	if len(existing) > 0 {
-		if _, ok := existing[0].(int32); ok {
-			marker = existing[0]
-		} else if _, ok := existing[0].(int); ok {
-			marker = existing[0]
-		}
-	}
-	var result bson.A
-	if marker != nil {
-		result = make(bson.A, 0, len(elements)+1)
-		result = append(result, marker)
-		result = append(result, elements...)
-	} else {
-		result = make(bson.A, len(elements))
-		copy(result, elements)
-	}
-	dSet(doc, key, result)
-}
-
-// extractBinaryIDFromDoc extracts a binary ID string from a bson.D field.
-func extractBinaryIDFromDoc(val any) string {
-	if bin, ok := val.(primitive.Binary); ok {
-		return types.BlobToUUID(bin.Data)
-	}
-	return ""
-}
-
-// ============================================================================
-// BSON widget tree walking (used by cmd_widgets.go)
-// ============================================================================
-
-// bsonWidgetResult holds a found widget and its parent context.
-type bsonWidgetResult struct {
-	widget      bson.D
-	parentArr   []any
-	parentKey   string
-	parentDoc   bson.D
-	index       int
-	colPropKeys map[string]string
-}
-
-// widgetFinder is a function type for locating widgets in a raw BSON tree.
-type widgetFinder func(rawData bson.D, widgetName string) *bsonWidgetResult
-
-// findBsonWidget searches the raw BSON page tree for a widget by name.
-func findBsonWidget(rawData bson.D, widgetName string) *bsonWidgetResult {
-	formCall := dGetDoc(rawData, "FormCall")
-	if formCall == nil {
-		return nil
-	}
-	args := dGetArrayElements(dGet(formCall, "Arguments"))
-	for _, arg := range args {
-		argDoc, ok := arg.(bson.D)
-		if !ok {
-			continue
-		}
-		if result := findInWidgetArray(argDoc, "Widgets", widgetName); result != nil {
-			return result
-		}
-	}
-	return nil
-}
-
-// findBsonWidgetInSnippet searches the raw BSON snippet tree for a widget by name.
-func findBsonWidgetInSnippet(rawData bson.D, widgetName string) *bsonWidgetResult {
-	if result := findInWidgetArray(rawData, "Widgets", widgetName); result != nil {
-		return result
-	}
-	if widgetContainer := dGetDoc(rawData, "Widget"); widgetContainer != nil {
-		if result := findInWidgetArray(widgetContainer, "Widgets", widgetName); result != nil {
-			return result
-		}
-	}
-	return nil
-}
-
-// findInWidgetArray searches a widget array (by key in parentDoc) for a named widget.
-func findInWidgetArray(parentDoc bson.D, key string, widgetName string) *bsonWidgetResult {
-	elements := dGetArrayElements(dGet(parentDoc, key))
-	for i, elem := range elements {
-		wDoc, ok := elem.(bson.D)
-		if !ok {
-			continue
-		}
-		if dGetString(wDoc, "Name") == widgetName {
-			return &bsonWidgetResult{
-				widget:    wDoc,
-				parentArr: elements,
-				parentKey: key,
-				parentDoc: parentDoc,
-				index:     i,
-			}
-		}
-		if result := findInWidgetChildren(wDoc, widgetName); result != nil {
-			return result
-		}
-	}
-	return nil
-}
-
-// findInWidgetChildren recursively searches widget children for a named widget.
-func findInWidgetChildren(wDoc bson.D, widgetName string) *bsonWidgetResult {
-	typeName := dGetString(wDoc, "$Type")
-
-	if result := findInWidgetArray(wDoc, "Widgets", widgetName); result != nil {
-		return result
-	}
-	if result := findInWidgetArray(wDoc, "FooterWidgets", widgetName); result != nil {
-		return result
-	}
-
-	// LayoutGrid: Rows[].Columns[].Widgets[]
-	if strings.Contains(typeName, "LayoutGrid") {
-		rows := dGetArrayElements(dGet(wDoc, "Rows"))
-		for _, row := range rows {
-			rowDoc, ok := row.(bson.D)
-			if !ok {
-				continue
-			}
-			cols := dGetArrayElements(dGet(rowDoc, "Columns"))
-			for _, col := range cols {
-				colDoc, ok := col.(bson.D)
-				if !ok {
-					continue
-				}
-				if result := findInWidgetArray(colDoc, "Widgets", widgetName); result != nil {
-					return result
-				}
-			}
-		}
-	}
-
-	// TabContainer: TabPages[].Widgets[]
-	tabPages := dGetArrayElements(dGet(wDoc, "TabPages"))
-	for _, tp := range tabPages {
-		tpDoc, ok := tp.(bson.D)
-		if !ok {
-			continue
-		}
-		if result := findInWidgetArray(tpDoc, "Widgets", widgetName); result != nil {
-			return result
-		}
-	}
-
-	// ControlBar
-	if controlBar := dGetDoc(wDoc, "ControlBar"); controlBar != nil {
-		if result := findInWidgetArray(controlBar, "Items", widgetName); result != nil {
-			return result
-		}
-	}
-
-	// CustomWidget (pluggable): Object.Properties[].Value.Widgets[]
-	if strings.Contains(typeName, "CustomWidget") {
-		if obj := dGetDoc(wDoc, "Object"); obj != nil {
-			props := dGetArrayElements(dGet(obj, "Properties"))
-			for _, prop := range props {
-				propDoc, ok := prop.(bson.D)
-				if !ok {
-					continue
-				}
-				if valDoc := dGetDoc(propDoc, "Value"); valDoc != nil {
-					if result := findInWidgetArray(valDoc, "Widgets", widgetName); result != nil {
-						return result
-					}
-				}
-			}
-		}
-	}
-
-	return nil
-}
-
-// setTranslatableText sets a translatable text value in BSON.
-func setTranslatableText(parent bson.D, key string, value interface{}) {
-	strVal, ok := value.(string)
-	if !ok {
-		return
-	}
-
-	target := parent
-	if key != "" {
-		if nested := dGetDoc(parent, key); nested != nil {
-			target = nested
-		} else {
-			dSet(parent, key, strVal)
-			return
-		}
-	}
-
-	translations := dGetArrayElements(dGet(target, "Translations"))
-	if len(translations) > 0 {
-		if tDoc, ok := translations[0].(bson.D); ok {
-			dSet(tDoc, "Text", strVal)
-			return
-		}
-	}
-	dSet(target, "Text", strVal)
-}
-
-// ============================================================================
-// Widget property setting (used by cmd_widgets.go)
-// ============================================================================
-
-// setRawWidgetProperty sets a property on a raw BSON widget document.
-func setRawWidgetProperty(widget bson.D, propName string, value interface{}) error {
-	switch propName {
-	case "Caption":
-		return setWidgetCaption(widget, value)
-	case "Content":
-		return setWidgetContent(widget, value)
-	case "Label":
-		return setWidgetLabel(widget, value)
-	case "ButtonStyle":
-		if s, ok := value.(string); ok {
-			dSet(widget, "ButtonStyle", s)
-		}
-		return nil
-	case "Class":
-		if appearance := dGetDoc(widget, "Appearance"); appearance != nil {
-			if s, ok := value.(string); ok {
-				dSet(appearance, "Class", s)
-			}
-		}
-		return nil
-	case "Style":
-		if appearance := dGetDoc(widget, "Appearance"); appearance != nil {
-			if s, ok := value.(string); ok {
-				dSet(appearance, "Style", s)
-			}
-		}
-		return nil
-	case "Editable":
-		if s, ok := value.(string); ok {
-			dSet(widget, "Editable", s)
-		}
-		return nil
-	case "Visible":
-		if s, ok := value.(string); ok {
-			dSet(widget, "Visible", s)
-		} else if b, ok := value.(bool); ok {
-			if b {
-				dSet(widget, "Visible", "True")
-			} else {
-				dSet(widget, "Visible", "False")
-			}
-		}
-		return nil
-	case "Name":
-		if s, ok := value.(string); ok {
-			dSet(widget, "Name", s)
-		}
-		return nil
-	case "Attribute":
-		return setWidgetAttributeRef(widget, value)
-	default:
-		return setPluggableWidgetProperty(widget, propName, value)
-	}
-}
-
-func setWidgetCaption(widget bson.D, value interface{}) error {
-	caption := dGetDoc(widget, "Caption")
-	if caption == nil {
-		setTranslatableText(widget, "Caption", value)
-		return nil
-	}
-	setTranslatableText(caption, "", value)
-	return nil
-}
-
-func setWidgetContent(widget bson.D, value interface{}) error {
-	strVal, ok := value.(string)
-	if !ok {
-		return mdlerrors.NewValidation("Content value must be a string")
-	}
-	content := dGetDoc(widget, "Content")
-	if content == nil {
-		return mdlerrors.NewValidation("widget has no Content property")
-	}
-	template := dGetDoc(content, "Template")
-	if template == nil {
-		return mdlerrors.NewValidation("Content has no Template")
-	}
-	items := dGetArrayElements(dGet(template, "Items"))
-	if len(items) > 0 {
-		if itemDoc, ok := items[0].(bson.D); ok {
-			dSet(itemDoc, "Text", strVal)
-			return nil
-		}
-	}
-	return mdlerrors.NewValidation("Content.Template has no Items with Text")
-}
-
-func setWidgetLabel(widget bson.D, value interface{}) error {
-	label := dGetDoc(widget, "Label")
-	if label == nil {
-		return nil
-	}
-	setTranslatableText(label, "Caption", value)
-	return nil
-}
-
-func setWidgetAttributeRef(widget bson.D, value interface{}) error {
-	attrPath, ok := value.(string)
-	if !ok {
-		return mdlerrors.NewValidation("Attribute value must be a string")
-	}
-
-	var attrRefValue interface{}
-	if strings.Count(attrPath, ".") >= 2 {
-		attrRefValue = bson.D{
-			{Key: "$ID", Value: bsonutil.NewIDBsonBinary()},
-			{Key: "$Type", Value: "DomainModels$AttributeRef"},
-			{Key: "Attribute", Value: attrPath},
-			{Key: "EntityRef", Value: nil},
-		}
-	} else {
-		attrRefValue = nil
-	}
-
-	for i, elem := range widget {
-		if elem.Key == "AttributeRef" {
-			widget[i].Value = attrRefValue
-			return nil
-		}
-	}
-	return mdlerrors.NewValidation("widget does not have an AttributeRef property; Attribute can only be SET on input widgets (TextBox, TextArea, DatePicker, etc.)")
-}
-
-func setPluggableWidgetProperty(widget bson.D, propName string, value interface{}) error {
-	obj := dGetDoc(widget, "Object")
-	if obj == nil {
-		return mdlerrors.NewNotFoundMsg("property", propName, fmt.Sprintf("property %q not found (widget has no pluggable Object)", propName))
-	}
-
-	propTypeKeyMap := make(map[string]string)
-	if widgetType := dGetDoc(widget, "Type"); widgetType != nil {
-		if objType := dGetDoc(widgetType, "ObjectType"); objType != nil {
-			propTypes := dGetArrayElements(dGet(objType, "PropertyTypes"))
-			for _, pt := range propTypes {
-				ptDoc, ok := pt.(bson.D)
-				if !ok {
-					continue
-				}
-				key := dGetString(ptDoc, "PropertyKey")
-				if key == "" {
-					continue
-				}
-				id := extractBinaryIDFromDoc(dGet(ptDoc, "$ID"))
-				if id != "" {
-					propTypeKeyMap[id] = key
-				}
-			}
-		}
-	}
-
-	props := dGetArrayElements(dGet(obj, "Properties"))
-	for _, prop := range props {
-		propDoc, ok := prop.(bson.D)
-		if !ok {
-			continue
-		}
-		typePointerID := extractBinaryIDFromDoc(dGet(propDoc, "TypePointer"))
-		propKey := propTypeKeyMap[typePointerID]
-		if propKey != propName {
-			continue
-		}
-		if valDoc := dGetDoc(propDoc, "Value"); valDoc != nil {
-			switch v := value.(type) {
-			case string:
-				dSet(valDoc, "PrimitiveValue", v)
-			case bool:
-				if v {
-					dSet(valDoc, "PrimitiveValue", "yes")
-				} else {
-					dSet(valDoc, "PrimitiveValue", "no")
-				}
-			case int:
-				dSet(valDoc, "PrimitiveValue", fmt.Sprintf("%d", v))
-			case float64:
-				dSet(valDoc, "PrimitiveValue", fmt.Sprintf("%g", v))
-			default:
-				dSet(valDoc, "PrimitiveValue", fmt.Sprintf("%v", v))
-			}
-			return nil
-		}
-		return mdlerrors.NewValidation(fmt.Sprintf("property %q has no Value map", propName))
-	}
-	return mdlerrors.NewNotFound("pluggable property", propName)
-}
