+++ mdl/executor/cmd_microflows_builder.go
@@ -8,9 +8,9 @@ import (
 	"strings"
 
 	"github.com/mendixlabs/mxcli/mdl/ast"
+	"github.com/mendixlabs/mxcli/mdl/backend"
 	"github.com/mendixlabs/mxcli/model"
 	"github.com/mendixlabs/mxcli/sdk/microflows"
-	"github.com/mendixlabs/mxcli/sdk/mpr"
 )
 
 // flowBuilder helps construct the flow graph from AST statements.
@@ -30,7 +30,7 @@ type flowBuilder struct {
 	measurer            *layoutMeasurer              // For measuring statement dimensions
 	nextConnectionPoint model.ID                     // For compound statements: the exit point differs from entry point
 	nextFlowCase        string                       // If set, next connecting flow uses this case value (for merge-less splits)
-	reader              *mpr.Reader                  // For looking up page/microflow references
+	reader              backend.FullBackend              // For looking up page/microflow references
 	hierarchy           *ContainerHierarchy          // For resolving container IDs to module names
 	pendingAnnotations  *ast.ActivityAnnotations     // Pending annotations to attach to next activity
 	restServices        []*model.ConsumedRestService // Cached REST services for parameter classification
